test(ses): reorganize the template and helper tests around their post-split owners
